Move recipe editor UI fragment routes and tests to ui_fragments module
