+++ tests/routers/test_ui_fragments.py
@@ -1,0 +1,1008 @@
+"""Tests for route handlers defined in meal_planner.routers.ui_fragments."""
+
+from typing import cast
+from unittest.mock import patch
+
+import monsterui.all as mu
+import pytest
+from bs4 import BeautifulSoup, Tag
+from httpx import AsyncClient
+from pydantic import ValidationError
+
+from meal_planner.models import RecipeBase
+from meal_planner.ui.common import CSS_ERROR_CLASS
+from tests.constants import (
+    FIELD_INGREDIENTS,
+    FIELD_INSTRUCTIONS,
+    FIELD_NAME,
+    FIELD_ORIGINAL_INGREDIENTS,
+    FIELD_ORIGINAL_INSTRUCTIONS,
+    FIELD_ORIGINAL_NAME,
+    FIELD_RECIPE_TEXT,  # For TestRecipeSortableListPersistence
+    RECIPES_EXTRACT_RUN_URL,  # For TestRecipeSortableListPersistence
+)
+from tests.test_helpers import extract_full_edit_form_data
+
+
+def _build_ui_fragment_form_data(
+    name="Test Recipe",
+    ingredients=None,
+    instructions=None,
+    original_name=None,
+    original_ingredients=None,
+    original_instructions=None,
+) -> dict:
+    ingredients = ingredients if ingredients is not None else ["ing1"]
+    instructions = instructions if instructions is not None else ["step1"]
+    original_name = original_name if original_name is not None else name
+    original_ingredients = (
+        original_ingredients if original_ingredients is not None else ingredients
+    )
+    original_instructions = (
+        original_instructions if original_instructions is not None else instructions
+    )
+
+    return {
+        "name": name,
+        "ingredients": ingredients,
+        "instructions": instructions,
+        "original_name": original_name,
+        "original_ingredients": original_ingredients,
+        "original_instructions": original_instructions,
+    }
+
+
+@pytest.mark.anyio
+class TestUpdateDiffEndpoint:
+    UPDATE_DIFF_URL = "/recipes/ui/update-diff"
+
+    def _build_diff_form_data(
+        self, current: RecipeBase, original: RecipeBase | None = None
+    ) -> dict:
+        if original is None:
+            original = current
+        form_data = {
+            FIELD_NAME: current.name,
+            FIELD_INGREDIENTS: current.ingredients,
+            FIELD_INSTRUCTIONS: current.instructions,
+            FIELD_ORIGINAL_NAME: original.name,
+            FIELD_ORIGINAL_INGREDIENTS: original.ingredients,
+            FIELD_ORIGINAL_INSTRUCTIONS: original.instructions,
+        }
+        return form_data
+
+    @patch("meal_planner.routers.ui_fragments.build_diff_content_children")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_diff_generation_error(
+        self, mock_logger_error, mock_build_diff, client: AsyncClient
+    ):
+        diff_exception = Exception("Simulated diff error")
+        mock_build_diff.side_effect = diff_exception
+        recipe = RecipeBase(name="Error Recipe", ingredients=["i"], instructions=["s"])
+        form_data = self._build_diff_form_data(recipe, recipe)
+
+        response = await client.post(self.UPDATE_DIFF_URL, data=form_data)
+        assert response.status_code == 200
+        html = response.text
+        assert f'class="{CSS_ERROR_CLASS}"' in html
+        assert "Error updating diff view" in html
+
+        mock_build_diff.assert_called_once()
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        assert args[0] == "Error updating diff: %s"
+        assert args[1] is diff_exception
+        assert kwargs.get("exc_info") is True
+
+    async def test_update_diff_happy_path(self, client: AsyncClient):
+        """Test successful diff generation and rendering."""
+        current = RecipeBase(
+            name="New Name", ingredients=["ing1", "new_ing"], instructions=["s1"]
+        )
+        original = RecipeBase(
+            name="Old Name", ingredients=["ing1"], instructions=["s1", "old_s2"]
+        )
+        form_data = self._build_diff_form_data(current, original)
+
+        response = await client.post(self.UPDATE_DIFF_URL, data=form_data)
+        assert response.status_code == 200
+        html = response.text
+        soup = BeautifulSoup(html, "html.parser")
+
+        wrapper_div = soup.find("div", id="diff-content-wrapper")
+        assert wrapper_div is not None
+        assert "flex" in wrapper_div.get("class", [])
+        assert "space-x-4" in wrapper_div.get("class", [])
+        assert "mt-4" in wrapper_div.get("class", [])
+
+        before_pre = wrapper_div.find("pre", id="diff-before-pre")
+        after_pre = wrapper_div.find("pre", id="diff-after-pre")
+        assert before_pre is not None
+        assert after_pre is not None
+
+        assert before_pre.find("del") is not None, "<del> tag missing in before_pre"
+        assert after_pre.find("ins") is not None, "<ins> tag missing in after_pre"
+        assert "Old Name" in before_pre.get_text()
+        assert "New Name" in after_pre.get_text()
+
+    @pytest.mark.parametrize(
+        "invalid_field, invalid_value, error_title_suffix",
+        [
+            pytest.param(FIELD_NAME, "", "curr_name", id="current_empty_name"),
+            pytest.param(
+                FIELD_INGREDIENTS, [""], "curr_ing", id="current_empty_ingredient"
+            ),
+            pytest.param(
+                FIELD_ORIGINAL_NAME, "", "orig_name", id="original_empty_name"
+            ),
+            pytest.param(
+                FIELD_ORIGINAL_INGREDIENTS,
+                [""],
+                "orig_ing",
+                id="original_empty_ingredient",
+            ),
+        ],
+    )
+    @patch("meal_planner.routers.ui_fragments.logger.warning")
+    async def test_update_diff_validation_error(
+        self,
+        mock_logger_warning,
+        client: AsyncClient,
+        invalid_field: str,
+        invalid_value: str | list[str],
+        error_title_suffix: str,
+    ):
+        valid_recipe = RecipeBase(name="Valid", ingredients=["i"], instructions=["s"])
+        form_data = self._build_diff_form_data(valid_recipe, valid_recipe)
+        form_data[invalid_field] = invalid_value
+
+        response = await client.post(self.UPDATE_DIFF_URL, data=form_data)
+
+        assert response.status_code == 200
+        html = response.text
+        assert "Recipe state invalid for diff. Please check all fields." in html
+        assert "diff-content-wrapper" not in html
+
+        mock_logger_warning.assert_called_once()
+        args, kwargs = mock_logger_warning.call_args
+        assert args[0] == "Validation error during diff update: %s"
+        assert isinstance(args[1], ValidationError)
+        assert kwargs.get("exc_info") is False
+
+    @pytest.mark.anyio
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_update_diff_parsing_exception(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        mock_parse.side_effect = Exception("Simulated parsing error")
+        dummy_form_data = {FIELD_NAME: "Test", "original_name": "Orig"}
+
+        response = await client.post(
+            TestUpdateDiffEndpoint.UPDATE_DIFF_URL, data=dummy_form_data
+        )
+
+        assert response.status_code == 200
+        assert "Error updating diff view." in response.text
+        assert CSS_ERROR_CLASS in response.text
+        assert mock_parse.call_count == 1
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        assert "Error updating diff: %s" in args[0]
+        assert isinstance(args[1], Exception)
+        assert args[1].args[0] == "Simulated parsing error"
+        assert kwargs.get("exc_info") is True
+
+
+@pytest.mark.anyio
+class TestRecipeSortableListPersistence:
+    INITIAL_RECIPE_TEXT = (
+        "Sortable Test Recipe\\\n"
+        "Ingredients: Ing1, Ing2, Ing3\\\n"
+        "Instructions: Step1, Step2"
+    )
+    MOCK_INITIAL_RECIPE = RecipeBase(
+        name="Sortable Test Recipe",
+        ingredients=["Ing1", "Ing2", "Ing3"],
+        instructions=["First instruction details", "Second instruction details"],
+    )
+
+    def _assert_sortable_attributes(
+        self, sortable_div: Tag | None, list_id_prefix: str
+    ):
+        assert sortable_div is not None, f"Sortable div '{list_id_prefix}' not found"
+        assert isinstance(sortable_div, Tag)
+        assert sortable_div.get("uk-sortable") == "handle: .drag-handle", (
+            f"uk-sortable missing or incorrect for {list_id_prefix}"
+        )
+        assert sortable_div.get("hx-trigger") == "moved", (
+            f"hx-trigger missing or incorrect for {list_id_prefix}"
+        )
+        assert sortable_div.get("hx-post") == "/recipes/ui/update-diff", (
+            f"hx-post missing or incorrect for {list_id_prefix}"
+        )
+        assert sortable_div.get("hx-target") == "#diff-content-wrapper", (
+            f"hx-target missing or incorrect for {list_id_prefix}"
+        )
+        assert sortable_div.get("hx-swap") == "innerHTML", (
+            f"hx-swap missing or incorrect for {list_id_prefix}"
+        )
+        assert sortable_div.get("hx-include") == "closest form", (
+            f"hx-include missing or incorrect for {list_id_prefix}"
+        )
+
+    @patch("meal_planner.main.generate_recipe_from_text")
+    async def test_sortable_after_ingredient_delete(
+        self, mock_llm_extract, client: AsyncClient
+    ):
+        mock_llm_extract.return_value = self.MOCK_INITIAL_RECIPE
+
+        extract_response = await client.post(
+            RECIPES_EXTRACT_RUN_URL, data={FIELD_RECIPE_TEXT: self.INITIAL_RECIPE_TEXT}
+        )
+        assert extract_response.status_code == 200
+        html_after_extract = extract_response.text
+
+        soup_after_extract = BeautifulSoup(html_after_extract, "html.parser")
+        edit_form_target_oob_div = soup_after_extract.find(
+            "div", attrs={"hx-swap-oob": "innerHTML:#edit-form-target"}
+        )
+        assert edit_form_target_oob_div is not None, (
+            "OOB div for edit form target not found after extract"
+        )
+        assert isinstance(edit_form_target_oob_div, Tag)
+
+        ingredients_list_div_extract = edit_form_target_oob_div.find(
+            "div", id="ingredients-list"
+        )
+        self._assert_sortable_attributes(
+            ingredients_list_div_extract, "ingredients-list (initial extract)"
+        )
+
+        form_data_for_delete = extract_full_edit_form_data(html_after_extract)
+
+        assert form_data_for_delete[FIELD_NAME] == self.MOCK_INITIAL_RECIPE.name
+        assert (
+            form_data_for_delete[FIELD_INGREDIENTS]
+            == self.MOCK_INITIAL_RECIPE.ingredients
+        )
+
+        index_to_delete = 1
+        delete_url = f"/recipes/ui/delete-ingredient/{index_to_delete}"
+        delete_response = await client.post(delete_url, data=form_data_for_delete)
+        assert delete_response.status_code == 200
+        html_after_delete = delete_response.text
+
+        soup_after_delete = BeautifulSoup(html_after_delete, "html.parser")
+        ingredients_list_div_after_delete = soup_after_delete.find(
+            "div", id="ingredients-list"
+        )
+        self._assert_sortable_attributes(
+            ingredients_list_div_after_delete, "ingredients-list (after delete)"
+        )
+
+        assert isinstance(ingredients_list_div_after_delete, Tag), (
+            "ingredients_list_div_after_delete is not a Tag"
+        )
+        inputs = ingredients_list_div_after_delete.find_all(
+            "input", attrs={"name": FIELD_INGREDIENTS}
+        )
+        assert len(inputs) == 2, f"Expected 2 ingredients, got {len(inputs)}"
+        assert inputs[0].get("value") == "Ing1"
+        assert inputs[1].get("value") == "Ing3"
+
+    @patch("meal_planner.main.generate_recipe_from_text")
+    async def test_sortable_after_instruction_delete(
+        self, mock_llm_extract, client: AsyncClient
+    ):
+        mock_llm_extract.return_value = self.MOCK_INITIAL_RECIPE
+
+        extract_response = await client.post(
+            RECIPES_EXTRACT_RUN_URL, data={FIELD_RECIPE_TEXT: self.INITIAL_RECIPE_TEXT}
+        )
+        assert extract_response.status_code == 200
+        html_after_extract = extract_response.text
+
+        soup_after_extract = BeautifulSoup(html_after_extract, "html.parser")
+        edit_form_target_oob_div = soup_after_extract.find(
+            "div", attrs={"hx-swap-oob": "innerHTML:#edit-form-target"}
+        )
+        assert edit_form_target_oob_div is not None
+        assert isinstance(edit_form_target_oob_div, Tag)
+
+        instructions_list_div_extract = edit_form_target_oob_div.find(
+            "div", id="instructions-list"
+        )
+        self._assert_sortable_attributes(
+            instructions_list_div_extract, "instructions-list (initial extract)"
+        )
+
+        form_data_for_delete = extract_full_edit_form_data(html_after_extract)
+        assert (
+            form_data_for_delete[FIELD_INSTRUCTIONS]
+            == self.MOCK_INITIAL_RECIPE.instructions
+        )
+
+        index_to_delete = 0
+        delete_url = f"/recipes/ui/delete-instruction/{index_to_delete}"
+        delete_response = await client.post(delete_url, data=form_data_for_delete)
+        assert delete_response.status_code == 200
+        html_after_delete = delete_response.text
+
+        soup_after_delete = BeautifulSoup(html_after_delete, "html.parser")
+        instructions_list_div_after_delete = soup_after_delete.find(
+            "div", id="instructions-list"
+        )
+        self._assert_sortable_attributes(
+            instructions_list_div_after_delete, "instructions-list (after delete)"
+        )
+
+        assert isinstance(instructions_list_div_after_delete, Tag)
+        textareas = instructions_list_div_after_delete.find_all(
+            "textarea", attrs={"name": FIELD_INSTRUCTIONS}
+        )
+        assert len(textareas) == 1, f"Expected 1 instruction, got {len(textareas)}"
+        assert (
+            textareas[0].get_text(strip=True) == "Second instruction details."
+        )  # Original was "Second instruction details"
+
+    @patch("meal_planner.main.generate_recipe_from_text")
+    async def test_sortable_after_ingredient_add(
+        self, mock_llm_extract, client: AsyncClient
+    ):
+        mock_llm_extract.return_value = self.MOCK_INITIAL_RECIPE
+
+        extract_response = await client.post(
+            RECIPES_EXTRACT_RUN_URL, data={FIELD_RECIPE_TEXT: self.INITIAL_RECIPE_TEXT}
+        )
+        assert extract_response.status_code == 200
+        html_after_extract = extract_response.text
+
+        soup_after_extract = BeautifulSoup(html_after_extract, "html.parser")
+        edit_form_target_oob_div = soup_after_extract.find(
+            "div", attrs={"hx-swap-oob": "innerHTML:#edit-form-target"}
+        )
+        assert edit_form_target_oob_div is not None
+        assert isinstance(edit_form_target_oob_div, Tag)
+        ingredients_list_div_extract = edit_form_target_oob_div.find(
+            "div", id="ingredients-list"
+        )
+        self._assert_sortable_attributes(
+            ingredients_list_div_extract, "ingredients-list (initial extract)"
+        )
+
+        form_data_for_add = extract_full_edit_form_data(html_after_extract)
+
+        add_url = "/recipes/ui/add-ingredient"
+        add_response = await client.post(add_url, data=form_data_for_add)
+        assert add_response.status_code == 200
+        html_after_add = add_response.text
+
+        soup_after_add = BeautifulSoup(html_after_add, "html.parser")
+        ingredients_list_div_after_add = soup_after_add.find(
+            "div", id="ingredients-list"
+        )
+        self._assert_sortable_attributes(
+            ingredients_list_div_after_add, "ingredients-list (after add)"
+        )
+
+        assert isinstance(ingredients_list_div_after_add, Tag)
+        inputs = ingredients_list_div_after_add.find_all(
+            "input", attrs={"name": FIELD_INGREDIENTS}
+        )
+        assert len(inputs) == 4, f"Expected 4 ingredients, got {len(inputs)}"
+        assert inputs[0].get("value") == "Ing1"
+        assert inputs[1].get("value") == "Ing2"
+        assert inputs[2].get("value") == "Ing3"
+        assert inputs[3].get("value", "") == ""
+
+    @patch("meal_planner.main.generate_recipe_from_text")
+    async def test_sortable_after_instruction_add(
+        self, mock_llm_extract, client: AsyncClient
+    ):
+        mock_llm_extract.return_value = self.MOCK_INITIAL_RECIPE
+
+        extract_response = await client.post(
+            RECIPES_EXTRACT_RUN_URL, data={FIELD_RECIPE_TEXT: self.INITIAL_RECIPE_TEXT}
+        )
+        assert extract_response.status_code == 200
+        html_after_extract = extract_response.text
+
+        soup_after_extract = BeautifulSoup(html_after_extract, "html.parser")
+        edit_form_target_oob_div = soup_after_extract.find(
+            "div", attrs={"hx-swap-oob": "innerHTML:#edit-form-target"}
+        )
+        assert edit_form_target_oob_div is not None
+        assert isinstance(edit_form_target_oob_div, Tag)
+        instructions_list_div_extract = edit_form_target_oob_div.find(
+            "div", id="instructions-list"
+        )
+        self._assert_sortable_attributes(
+            instructions_list_div_extract, "instructions-list (initial extract)"
+        )
+
+        form_data_for_add = extract_full_edit_form_data(html_after_extract)
+
+        add_url = "/recipes/ui/add-instruction"
+        add_response = await client.post(add_url, data=form_data_for_add)
+        assert add_response.status_code == 200
+        html_after_add = add_response.text
+
+        soup_after_add = BeautifulSoup(html_after_add, "html.parser")
+        instructions_list_div_after_add = soup_after_add.find(
+            "div", id="instructions-list"
+        )
+        self._assert_sortable_attributes(
+            instructions_list_div_after_add, "instructions-list (after add)"
+        )
+
+        assert isinstance(instructions_list_div_after_add, Tag)
+        textareas = instructions_list_div_after_add.find_all(
+            "textarea", attrs={"name": FIELD_INSTRUCTIONS}
+        )
+        assert len(textareas) == 3, f"Expected 3 instructions, got {len(textareas)}"
+        assert (
+            textareas[0].get_text(strip=True) == "First instruction details."
+        )  # Original was "First instruction details"
+        assert (
+            textareas[1].get_text(strip=True) == "Second instruction details."
+        )  # Original was "Second instruction details"
+        assert textareas[2].get_text(strip=True) == ""
+
+
+@pytest.mark.anyio
+class TestRecipeUIFragments:
+    ADD_INGREDIENT_URL = "/recipes/ui/add-ingredient"
+    ADD_INSTRUCTION_URL = "/recipes/ui/add-instruction"
+    DELETE_INGREDIENT_BASE_URL = "/recipes/ui/delete-ingredient"
+    DELETE_INSTRUCTION_BASE_URL = "/recipes/ui/delete-instruction"
+
+    async def test_add_ingredient(self, client: AsyncClient):
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["existing ing"], instructions=["step1"]
+        )
+        response = await client.post(self.ADD_INGREDIENT_URL, data=form_data)
+        assert response.status_code == 200
+        soup = BeautifulSoup(response.text, "html.parser")
+
+        ingredients_list_div = soup.find("div", id="ingredients-list")
+        assert ingredients_list_div, "Ingredients list div not found"
+        assert isinstance(ingredients_list_div, Tag)
+        ingredient_inputs = ingredients_list_div.find_all(
+            "input", {"name": "ingredients"}
+        )
+        assert len(ingredient_inputs) == 2, (
+            f"Expected 2 ingredient inputs, got {len(ingredient_inputs)}"
+        )
+
+        new_ingredient_input = ingredient_inputs[-1]
+        assert isinstance(new_ingredient_input, Tag)
+        assert new_ingredient_input.get("value", "") == ""
+        assert new_ingredient_input["placeholder"] == "Ingredient"
+
+        new_item_div = new_ingredient_input.find_parent("div", class_="flex")
+        assert new_item_div, "Parent div for new ingredient not found"
+        assert isinstance(new_item_div, Tag)
+        delete_button = new_item_div.find(
+            "button",
+            {
+                "hx-post": lambda x: bool(
+                    x and x.startswith(f"{self.DELETE_INGREDIENT_BASE_URL}/")
+                )
+            },
+        )
+        assert delete_button, "Delete button for new ingredient not found"
+        assert isinstance(delete_button, Tag)
+
+        icon_element = delete_button.find("uk-icon", attrs={"icon": "minus-circle"})
+        assert icon_element, (
+            "UkIcon 'minus-circle' not found in ingredient delete button"
+        )
+        assert isinstance(icon_element, Tag)
+        class_list = icon_element.get("class")
+        if class_list is None:
+            class_list = []
+        assert str(mu.TextT.error) in class_list
+
+        oob_div = soup.find("div", {"hx-swap-oob": "innerHTML:#diff-content-wrapper"})
+        assert oob_div, "OOB diff wrapper not found"
+        assert isinstance(oob_div, Tag)
+        assert oob_div.find("pre", id="diff-before-pre"), "Diff before pre not found"
+        assert oob_div.find("pre", id="diff-after-pre"), "Diff after pre not found"
+
+    async def test_add_instruction(self, client: AsyncClient):
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=["existing step"]
+        )
+        response = await client.post(self.ADD_INSTRUCTION_URL, data=form_data)
+        assert response.status_code == 200
+        soup = BeautifulSoup(response.text, "html.parser")
+
+        instructions_list_div = soup.find("div", id="instructions-list")
+        assert instructions_list_div, "Instructions list div not found"
+        assert isinstance(instructions_list_div, Tag)
+        instruction_textareas = instructions_list_div.find_all(
+            "textarea", {"name": "instructions"}
+        )
+        assert len(instruction_textareas) == 2, (
+            f"Expected 2 instruction textareas, got {len(instruction_textareas)}"
+        )
+
+        new_instruction_textarea = instruction_textareas[-1]
+        assert isinstance(new_instruction_textarea, Tag)
+        assert new_instruction_textarea.text == ""
+        assert new_instruction_textarea["placeholder"] == "Instruction Step"
+
+        new_item_div = new_instruction_textarea.find_parent("div", class_="flex")
+        assert new_item_div, "Parent div for new instruction not found"
+        assert isinstance(new_item_div, Tag)
+        delete_button = new_item_div.find(
+            "button",
+            {
+                "hx-post": lambda x: bool(
+                    x and x.startswith(f"{self.DELETE_INSTRUCTION_BASE_URL}/")
+                )
+            },
+        )
+        assert delete_button, "Delete button for new instruction not found"
+        assert isinstance(delete_button, Tag)
+        icon_element = delete_button.find("uk-icon", attrs={"icon": "minus-circle"})
+        assert icon_element, (
+            "UkIcon 'minus-circle' not found in instruction delete button"
+        )
+        assert isinstance(icon_element, Tag)
+        class_list = icon_element.get("class")
+        if class_list is None:
+            class_list = []
+        assert str(mu.TextT.error) in class_list
+
+        oob_div = soup.find("div", {"hx-swap-oob": "innerHTML:#diff-content-wrapper"})
+        assert oob_div, "OOB diff wrapper not found"
+        assert isinstance(oob_div, Tag)
+        assert oob_div.find("pre", id="diff-before-pre"), "Diff before pre not found"
+        assert oob_div.find("pre", id="diff-after-pre"), "Diff after pre not found"
+
+    async def test_delete_ingredient(self, client: AsyncClient):
+        initial_ingredients = ["ing_to_keep1", "ing_to_delete", "ing_to_keep2"]
+        form_data = _build_ui_fragment_form_data(
+            ingredients=initial_ingredients, instructions=["step1"]
+        )
+        index_to_delete = 1
+
+        response = await client.post(
+            f"{self.DELETE_INGREDIENT_BASE_URL}/{index_to_delete}", data=form_data
+        )
+        assert response.status_code == 200
+        soup = BeautifulSoup(response.text, "html.parser")
+
+        ingredients_list_div = soup.find("div", id="ingredients-list")
+        assert ingredients_list_div, "Ingredients list div for delete not found"
+        assert isinstance(ingredients_list_div, Tag)
+        ingredient_inputs = ingredients_list_div.find_all(
+            "input", {"name": "ingredients"}
+        )
+        assert len(ingredient_inputs) == 2, (
+            f"Expected 2 ingredients after delete, got {len(ingredient_inputs)}"
+        )
+
+        rendered_ingredient_values = [
+            cast(Tag, inp)["value"] for inp in ingredient_inputs
+        ]
+        assert "ing_to_keep1" in rendered_ingredient_values
+        assert "ing_to_keep2" in rendered_ingredient_values
+        assert "ing_to_delete" not in rendered_ingredient_values
+
+        oob_div = soup.find("div", {"hx-swap-oob": "innerHTML:#diff-content-wrapper"})
+        assert oob_div, "OOB diff wrapper for delete not found"
+        assert isinstance(oob_div, Tag)
+        assert oob_div.find("pre", id="diff-before-pre"), (
+            "Diff before pre for delete not found"
+        )
+        assert oob_div.find("pre", id="diff-after-pre"), (
+            "Diff after pre for delete not found"
+        )
+
+    async def test_delete_instruction(self, client: AsyncClient):
+        initial_instructions = ["step_to_keep1", "step_to_delete", "step_to_keep2"]
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=initial_instructions
+        )
+        index_to_delete = 1
+
+        response = await client.post(
+            f"{self.DELETE_INSTRUCTION_BASE_URL}/{index_to_delete}", data=form_data
+        )
+        assert response.status_code == 200
+        soup = BeautifulSoup(response.text, "html.parser")
+
+        instructions_list_div = soup.find("div", id="instructions-list")
+        assert instructions_list_div, "Instructions list div for delete not found"
+        assert isinstance(instructions_list_div, Tag)
+        instruction_textareas = instructions_list_div.find_all(
+            "textarea", {"name": "instructions"}
+        )
+        assert len(instruction_textareas) == 2, (
+            f"Expected 2 instructions after delete, got {len(instruction_textareas)}"
+        )
+
+        rendered_instruction_values = [ta.text for ta in instruction_textareas]
+        assert "step_to_keep1" in rendered_instruction_values
+        assert "step_to_keep2" in rendered_instruction_values
+        assert "step_to_delete" not in rendered_instruction_values
+
+        oob_div = soup.find("div", {"hx-swap-oob": "innerHTML:#diff-content-wrapper"})
+        assert oob_div, "OOB diff wrapper for delete not found"
+        assert isinstance(oob_div, Tag)
+        assert oob_div.find("pre", id="diff-before-pre"), (
+            "Diff before pre for delete not found"
+        )
+        assert oob_div.find("pre", id="diff-after-pre"), (
+            "Diff after pre for delete not found"
+        )
+
+    @pytest.mark.parametrize("invalid_index", [5])
+    @patch("meal_planner.routers.ui_fragments.logger.warning")
+    async def test_delete_ingredient_invalid_index(
+        self, mock_logger_warning, client: AsyncClient, invalid_index: int
+    ):
+        initial_ingredients = ["ing1", "ing2"]
+        form_data = _build_ui_fragment_form_data(
+            ingredients=initial_ingredients, instructions=["step1"]
+        )
+
+        response = await client.post(
+            f"{self.DELETE_INGREDIENT_BASE_URL}/{invalid_index}", data=form_data
+        )
+        assert response.status_code == 200
+        soup = BeautifulSoup(response.text, "html.parser")
+        ingredient_inputs = soup.find_all("input", {"name": "ingredients"})
+        assert len(ingredient_inputs) == len(initial_ingredients)
+        mock_logger_warning.assert_called_once_with(
+            f"Attempted to delete ingredient at invalid index {invalid_index}"
+        )
+
+    @pytest.mark.parametrize("invalid_index", [5])
+    @patch("meal_planner.routers.ui_fragments.logger.warning")
+    async def test_delete_instruction_invalid_index(
+        self, mock_logger_warning, client: AsyncClient, invalid_index: int
+    ):
+        initial_instructions = ["inst1", "inst2"]
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=initial_instructions
+        )
+
+        response = await client.post(
+            f"{self.DELETE_INSTRUCTION_BASE_URL}/{invalid_index}", data=form_data
+        )
+        assert response.status_code == 200
+        soup = BeautifulSoup(response.text, "html.parser")
+        instruction_textareas = soup.find_all("textarea", {"name": "instructions"})
+        assert len(instruction_textareas) == len(initial_instructions)
+        mock_logger_warning.assert_called_once_with(
+            f"Attempted to delete instruction at invalid index {invalid_index}"
+        )
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_delete_ingredient_validation_error(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        validation_exc = ValidationError.from_exception_data(
+            title="TestVE_del_ing", line_errors=[]
+        )
+        initial_ingredients = ["ing1_to_delete", "ing2_remains"]
+        form_data_dict = _build_ui_fragment_form_data(
+            ingredients=initial_ingredients, instructions=["s1"]
+        )
+
+        mock_parse.side_effect = [
+            validation_exc,  # First call to _parse_recipe_form_data in the try block
+            form_data_dict.copy(),  # Second call in the except ValidationError block
+        ]
+        index_to_delete = 0
+
+        response = await client.post(
+            f"{self.DELETE_INGREDIENT_BASE_URL}/{index_to_delete}", data=form_data_dict
+        )
+        assert response.status_code == 200
+        assert "Error updating list after delete. Validation failed." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg_fragment = "Validation error processing ingredient deletion"
+        assert expected_log_msg_fragment in args[0]
+        # Ensure the specific exception 'e' is part of the log message
+        assert str(validation_exc) in args[0]  # Check if the specific error is logged
+        assert kwargs.get("exc_info") is True
+
+        # Check that the list is re-rendered with original items on validation error
+        soup = BeautifulSoup(response.text, "html.parser")
+        ingredients_list_div = soup.find("div", id="ingredients-list")
+        assert ingredients_list_div, "Ingredients list div for error case not found"
+        assert isinstance(ingredients_list_div, Tag)
+        ingredient_inputs = ingredients_list_div.find_all(
+            "input", {"name": "ingredients"}
+        )
+        assert len(ingredient_inputs) == len(initial_ingredients)
+
+        # Check that OOB diff is NOT sent
+        oob_div = soup.find("div", {"hx-swap-oob": "innerHTML:#diff-content-wrapper"})
+        assert not oob_div, "OOB diff wrapper should NOT be present on validation error"
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_delete_ingredient_generic_exception(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        generic_exc = Exception("Generic parse error for del_ing")
+        mock_parse.side_effect = generic_exc  # This will be raised by the first call
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=["s1"]
+        )
+        index_to_delete = 0
+
+        response = await client.post(
+            f"{self.DELETE_INGREDIENT_BASE_URL}/{index_to_delete}", data=form_data
+        )
+        assert response.status_code == 200
+        assert "Error processing delete request." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg = (
+            f"Error deleting ingredient at index {index_to_delete}: {generic_exc}"
+        )
+        assert args[0] == expected_log_msg
+        assert kwargs.get("exc_info") is True
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_delete_instruction_validation_error(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        validation_exc = ValidationError.from_exception_data(
+            title="TestVE_del_inst", line_errors=[]
+        )
+        initial_instructions = ["s1_to_delete", "s2_remains"]
+        form_data_dict = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=initial_instructions
+        )
+
+        mock_parse.side_effect = [
+            validation_exc,  # First call
+            form_data_dict.copy(),  # Second call in except
+        ]
+        index_to_delete = 0
+
+        response = await client.post(
+            f"{self.DELETE_INSTRUCTION_BASE_URL}/{index_to_delete}", data=form_data_dict
+        )
+        assert response.status_code == 200
+        assert "Error updating list after delete. Validation failed." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg_fragment = "Validation error processing instruction deletion"
+        assert expected_log_msg_fragment in args[0]
+        assert str(validation_exc) in args[0]
+        assert kwargs.get("exc_info") is True
+
+        soup = BeautifulSoup(response.text, "html.parser")
+        instructions_list_div = soup.find("div", id="instructions-list")
+        assert instructions_list_div, "Instructions list div for error case not found"
+        assert isinstance(instructions_list_div, Tag)
+        instruction_textareas = instructions_list_div.find_all(
+            "textarea", {"name": "instructions"}
+        )
+        assert len(instruction_textareas) == len(initial_instructions)
+        oob_div = soup.find("div", {"hx-swap-oob": "innerHTML:#diff-content-wrapper"})
+        assert not oob_div, "OOB diff wrapper should NOT be present on validation error"
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_delete_instruction_generic_exception(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        generic_exc = Exception("Generic parse error for del_inst")
+        mock_parse.side_effect = generic_exc
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=["s1"]
+        )
+        index_to_delete = 0
+
+        response = await client.post(
+            f"{self.DELETE_INSTRUCTION_BASE_URL}/{index_to_delete}", data=form_data
+        )
+        assert response.status_code == 200
+        assert "Error processing delete request." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg = (
+            f"Error deleting instruction at index {index_to_delete}: {generic_exc}"
+        )
+        assert args[0] == expected_log_msg
+        assert kwargs.get("exc_info") is True
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_add_ingredient_validation_error(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        validation_exc = ValidationError.from_exception_data(
+            title="TestVE_add_ing", line_errors=[]
+        )
+        # Simulate _parse_recipe_form_data in the except block
+        # It will be called with the original form_data if the try block fails
+        original_form_data_parsed = _build_ui_fragment_form_data(ingredients=["ing1"])
+
+        mock_parse.side_effect = [
+            validation_exc,  # First call in try block
+            original_form_data_parsed,  # Second call in except ValidationError block
+        ]
+        form_data = (
+            _build_ui_fragment_form_data(  # This is the form_data passed to the route
+                ingredients=["ing1"], instructions=["s1"]
+            )
+        )
+
+        response = await client.post(self.ADD_INGREDIENT_URL, data=form_data)
+        assert response.status_code == 200
+        assert "Error updating list after add." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg = (
+            f"Validation error processing ingredient addition: {validation_exc}"
+        )
+        assert args[0] == expected_log_msg
+        assert kwargs.get("exc_info") is True
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_add_ingredient_generic_exception(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        generic_exc = Exception("Generic parse error for add_ing")
+        mock_parse.side_effect = generic_exc
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=["s1"]
+        )
+
+        response = await client.post(self.ADD_INGREDIENT_URL, data=form_data)
+        assert response.status_code == 200
+        assert "Error processing add request." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg = f"Error adding ingredient: {generic_exc}"
+        assert args[0] == expected_log_msg
+        assert kwargs.get("exc_info") is True
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_add_instruction_validation_error(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        validation_exc = ValidationError.from_exception_data(
+            title="TestVE_add_inst", line_errors=[]
+        )
+        original_form_data_parsed = _build_ui_fragment_form_data(instructions=["s1"])
+        mock_parse.side_effect = [
+            validation_exc,  # First call in try block
+            original_form_data_parsed,  # Second call in except ValidationError block
+        ]
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=["s1"]
+        )
+
+        response = await client.post(self.ADD_INSTRUCTION_URL, data=form_data)
+        assert response.status_code == 200
+        assert "Error updating list after add." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg = (
+            f"Validation error processing instruction addition: {validation_exc}"
+        )
+        assert args[0] == expected_log_msg
+        assert kwargs.get("exc_info") is True
+
+    @patch("meal_planner.routers.ui_fragments._parse_recipe_form_data")
+    @patch("meal_planner.routers.ui_fragments.logger.error")
+    async def test_add_instruction_generic_exception(
+        self, mock_logger_error, mock_parse, client: AsyncClient
+    ):
+        generic_exc = Exception("Generic parse error for add_inst")
+        mock_parse.side_effect = generic_exc
+        form_data = _build_ui_fragment_form_data(
+            ingredients=["ing1"], instructions=["s1"]
+        )
+
+        response = await client.post(self.ADD_INSTRUCTION_URL, data=form_data)
+        assert response.status_code == 200
+        assert "Error processing add request." in response.text
+        assert CSS_ERROR_CLASS in response.text
+
+        mock_logger_error.assert_called_once()
+        args, kwargs = mock_logger_error.call_args
+        expected_log_msg = f"Error adding instruction: {generic_exc}"
+        assert args[0] == expected_log_msg
+        assert kwargs.get("exc_info") is True
+
+    async def test_delete_instruction_missing_ingredients_in_form(
+        self, client: AsyncClient
+    ):
+        # This test checks if the _parse_recipe_form_data in the except ValidationError
+        # block handles missing 'ingredients' gracefully (it should default to []).
+        initial_instructions = ["step_to_keep1", "step_to_delete", "step_to_keep2"]
+        index_to_delete = 1
+
+        # Form data intentionally missing 'ingredients' and 'original_ingredients'
+        form_data_dict = {
+            "name": "Test Recipe Name",
+            "instructions": initial_instructions,
+            "original_name": "Test Recipe Name",
+            "original_instructions": initial_instructions,
+            # 'ingredients' and 'original_ingredients' are missing
+        }
+
+        # To trigger the ValidationError path in post_delete_instruction_row,
+        # we need _parse_recipe_form_data (the first call in the `try` block)
+        # to raise a ValidationError.
+        # We'll mock _parse_recipe_form_data to do this for the first call,
+        # and for the second call (in the `except` block), it will parse
+        # the incomplete form_data_dict.
+        with patch(
+            "meal_planner.routers.ui_fragments._parse_recipe_form_data"
+        ) as mock_parse_dynamic:
+            validation_error = ValidationError.from_exception_data(
+                title="mock VE", line_errors=[]
+            )
+            # First call raises VE, second call processes the (incomplete) form_data
+            mock_parse_dynamic.side_effect = [
+                validation_error,
+                _build_ui_fragment_form_data(
+                    name=form_data_dict["name"],
+                    ingredients=[],  # This is what _parse_recipe_form_data would do
+                    instructions=form_data_dict["instructions"],
+                    original_name=form_data_dict["original_name"],
+                    original_ingredients=[],  # Parsed as empty list if missing
+                    original_instructions=form_data_dict["original_instructions"],
+                ),
+            ]
+
+            response = await client.post(
+                f"{self.DELETE_INSTRUCTION_BASE_URL}/{index_to_delete}",
+                data=form_data_dict,
+            )
+        assert response.status_code == 200
+        soup = BeautifulSoup(response.text, "html.parser")
+
+        assert "Error updating list after delete. Validation failed." in response.text
+
+        instructions_list_div = soup.find("div", id="instructions-list")
+        assert instructions_list_div, "Instructions list div for delete error not found"
+        assert isinstance(instructions_list_div, Tag)
+        instruction_textareas = instructions_list_div.find_all(
+            "textarea", {"name": "instructions"}
+        )
+        # It should re-render the original list of instructions
+        assert len(instruction_textareas) == len(initial_instructions)
+
+        oob_div = soup.find("div", {"hx-swap-oob": "innerHTML:#diff-content-wrapper"})
+        assert not oob_div, (
+            "OOB diff wrapper should NOT be present on validation error path"
+        )
+
+
+# Tests for generate_diff_html and build_edit_review_form are specific to ui.edit_recipe
+# and should remain in tests/test_ui/test_recipe_editor.py, not moved here.
+# This file is for testing the route handlers in ui_fragments.py.
